LayoutTests/webgpu/command-buffers.html

<!DOCTYPE html>
<html>
<script src="js/webgpu-functions.js"></script>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>
promise_test(async () => {
    const device = await getBasicDevice();
    const commandBuffer = device.createCommandBuffer();
    assert_true(commandBuffer instanceof WebGPUCommandBuffer, "Successfully created GPUCommandBuffer.");
}, "Create a default GPUCommandBuffer.");

// FIXME: createCommandBuffer should take a GPUCommandBufferDescriptor, which is currently an empty dictionary.
</script>
</html>
